Alteração nas implementações
Processo seletivo e participante é onde há mais trabalho. Afinal houveram alguns erros quanto a lógica da implementação e o modelo real que o projeto deve seguir.

+++ prodigialis_core/src/br/cefetmg/inf/prodigialis/model/dao/impl/ProvaDAO.java
@@ -164,7 +164,7 @@ public class ProvaDAO implements IProvaDAO{
             while(resultSet.next()){
                 
                 Prova prova = new Prova();
-                prova.setCod_prova(resultSet.getLong("cod_prova"));
+                prova.setCod_prova(resultSet.getInt("cod_prova"));
                 EmpresaProva empresaProva = empresaProvaDAO.consultarPorId(resultSet.getString("cnpj").charAt(0));
                 prova.setEmpresa(empresaProva);
                 prova.setDat_prova(resultSet.getDate("dat_prova"));
@@ -212,7 +212,7 @@ public class ProvaDAO implements IProvaDAO{
             while(resultSet.next()){
                 
                 prova = new Prova();
-                prova.setCod_prova(resultSet.getLong("cod_prova"));
+                prova.setCod_prova(resultSet.getInt("cod_prova"));
                 EmpresaProva empresaProva = empresaProvaDAO.consultarPorId(resultSet.getString("cnpj").charAt(0));
                 prova.setEmpresa(empresaProva);
                 prova.setDat_prova(resultSet.getDate("dat_prova"));
